+++ cmd/run.go
@@ -39,6 +39,7 @@ func newRunCommand() *cobra.Command {
 	broker.KafkaAddresses(runCmd.Flags())
 	broker.KafkaGroup(runCmd.Flags())
 	broker.KafkaTopicTxSender(runCmd.Flags())
+	broker.KafkaTopicTxRecover(runCmd.Flags())
 
 	// Register StoreGRPC flags
 	storeclient.EnvelopeStoreGRPCTarget(runCmd.Flags())