+++ handlers/producer/producer.go
@@ -1,0 +1,40 @@
+package producer
+
+import (
+	"github.com/Shopify/sarama"
+	"github.com/spf13/viper"
+	encoding "gitlab.com/ConsenSys/client/fr/core-stack/pkg.git/encoding/sarama"
+	"gitlab.com/ConsenSys/client/fr/core-stack/pkg.git/engine"
+	"gitlab.com/ConsenSys/client/fr/core-stack/pkg.git/errors"
+	"gitlab.com/ConsenSys/client/fr/core-stack/pkg.git/handlers/producer"
+	"gitlab.com/ConsenSys/client/fr/core-stack/pkg.git/utils"
+)
+
+// PrepareMsg prepare message to produce from TxContexts
+func PrepareMsg(txctx *engine.TxContext, msg *sarama.ProducerMessage) error {
+	// If an error occurred then we redirect to recovery
+	for _, err := range txctx.Envelope.GetErrors() {
+		if errors.IsWarning(err) {
+			continue
+		}
+
+		// Marshal Envelope into sarama Message
+		err := encoding.Marshal(txctx.Envelope, msg)
+		if err != nil {
+			return err
+		}
+
+		// Set Topic to Nonce topic
+		msg.Topic = viper.GetString("kafka.topic.recover")
+		msg.Key = sarama.StringEncoder(utils.ToChainAccountKey(txctx.Envelope.GetChain().ID(), txctx.Envelope.Sender()))
+
+		break
+	}
+
+	return nil
+}
+
+// Producer creates a producer handler
+func Producer(p sarama.SyncProducer) engine.HandlerFunc {
+	return producer.Producer(p, PrepareMsg)
+}